requirements.txt
streamlit
google-generativeai
PyMuPDF
scikit-learn
numpy
